refactor: extract upload file selection
