+++ src/renderer/app/directives/add-torrent-modal/add-torrent-modal.controller.ts
@@ -1,26 +1,13 @@
-import { ITimeoutService } from "angular";
 import { TorrentUploadOptions } from "@renderer/app/bittorrent/torrentclient";
 import { ModalController } from "@renderer/app/directives/modal/modal.controller";
 import { SavedLocationModalController } from "@renderer/app/directives/saved-location-modal/saved-location-modal.controller";
 import type { ElectorrentRootScope } from "@renderer/app/types/root-scope";
 import { AddTorrentModalScope } from "./add-torrent-modal.directive";
-import type { BittorrentFileSelection, TorrentMetadataFile } from "@shared/ipc-contract";
-
-interface UploadFileTreeNode {
-    id: string
-    name: string
-    path: string
-    size: number
-    level: number
-    fileIndex?: number
-    children: UploadFileTreeNode[]
-    selected: boolean
-    partial: boolean
-}
+import type { BittorrentFileSelection } from "@shared/ipc-contract";
 
 export class AddTorrentModalController {
 
-    static $inject = ["$scope", "$rootScope", "$timeout"]
+    static $inject = ["$scope", "$rootScope"]
 
     static defaultTorrentUploadOptions: TorrentUploadOptions = {
         startTorrent: true,
@@ -32,13 +19,11 @@ export class AddTorrentModalController {
     savedLocationModalRef: SavedLocationModalController
     uploadOptions: TorrentUploadOptions
     isLoading: boolean
-    activeTab = "general"
-    fileTree: UploadFileTreeNode[] = []
-    visibleFileTree: UploadFileTreeNode[] = []
+    activeTab: "general" | "files" = "general"
     private preserveUploadsOnHide: boolean
     private restoreUploadOptionsOnShow: boolean
 
-    constructor(scope: AddTorrentModalScope, rootScope: ElectorrentRootScope, private readonly $timeout: ITimeoutService) {
+    constructor(scope: AddTorrentModalScope, rootScope: ElectorrentRootScope) {
         this.scope = scope
         this.rootScope = rootScope
         this.isLoading = false
@@ -49,7 +34,6 @@ export class AddTorrentModalController {
             return this.scope.torrents && this.scope.torrents.length
         }, (newVal, oldVal) => {
             if (newVal > 0) {
-                this.refreshUploadFiles()
                 this.modalref.showModal()
             } else if (oldVal > 0 && newVal === 0) {
                 this.modalref.hideModal()
@@ -74,7 +58,6 @@ export class AddTorrentModalController {
             configuredOptions || {},
         )
         this.activeTab = "general"
-        this.refreshUploadFiles()
     }
 
     onHidden() {
@@ -82,8 +65,6 @@ export class AddTorrentModalController {
             return
         }
         this.scope.torrents = []
-        this.fileTree = []
-        this.visibleFileTree = []
     }
 
     supportsSavedLocations() {
@@ -111,8 +92,13 @@ export class AddTorrentModalController {
         return metadata?.length || metadata?.files.reduce((size, file) => size + (file.length || 0), 0) || 0
     }
 
+    getCurrentTorrentUploadFiles() {
+        return this.getCurrentTorrentUpload()?.metadata?.files
+    }
+
     hasFilesTab() {
-        return !!this.rootScope.$btclient?.features.uploadFileSelection && this.visibleFileTree.length > 0
+        return !!this.rootScope.$btclient?.features.uploadFileSelection
+            && !!this.getCurrentTorrentUploadFiles()?.length
     }
 
     switchTab(tab: "general" | "files") {
@@ -120,13 +106,14 @@ export class AddTorrentModalController {
             return
         }
         this.activeTab = tab
-        this.updateIndeterminateCheckboxes()
     }
 
-    toggleFileNode(node: UploadFileTreeNode) {
-        this.setNodeSelection(node, node.partial ? true : !node.selected)
-        this.syncFileSelectionOptions()
-        this.updateIndeterminateCheckboxes()
+    updateFileSelection(selection?: BittorrentFileSelection[]) {
+        if (selection) {
+            this.uploadOptions.fileSelection = selection
+        } else {
+            delete this.uploadOptions.fileSelection
+        }
     }
 
     getPendingUploadCountLabel() {
@@ -199,135 +186,6 @@ export class AddTorrentModalController {
         } else {
             await this.rootScope.$btclient.uploadTorrent(torrent, filename, options, sourcePath)
         }
-    }
-
-    private refreshUploadFiles() {
-        const torrent = this.getCurrentTorrentUpload()
-        const files = torrent?.metadata?.files || []
-        if (!this.rootScope.$btclient?.features.uploadFileSelection || files.length === 0) {
-            delete this.uploadOptions.fileSelection
-            this.fileTree = []
-            this.visibleFileTree = []
-            this.activeTab = "general"
-            return
-        }
-
-        this.fileTree = this.buildFileTree(files)
-        this.visibleFileTree = this.flattenFileTree(this.fileTree)
-        this.syncFileSelectionOptions()
-        this.updateIndeterminateCheckboxes()
-    }
-
-    private buildFileTree(files: TorrentMetadataFile[]) {
-        const roots: UploadFileTreeNode[] = []
-        const folders = new Map<string, UploadFileTreeNode>()
-
-        files.forEach((file, index) => {
-            const normalizedPath = file.path || file.name || `file-${index + 1}`
-            const parts = normalizedPath.split(/[\\/]+/).filter(Boolean)
-            let siblings = roots
-            let folderPath = ""
-
-            parts.forEach((part, partIndex) => {
-                const isFile = partIndex === parts.length - 1
-                folderPath = folderPath ? `${folderPath}/${part}` : part
-                if (isFile) {
-                    siblings.push({
-                        id: `file-${index}`,
-                        name: part,
-                        path: normalizedPath,
-                        size: file.length || 0,
-                        level: partIndex,
-                        fileIndex: index,
-                        children: [],
-                        selected: true,
-                        partial: false,
-                    })
-                    return
-                }
-
-                let folder = folders.get(folderPath)
-                if (!folder) {
-                    folder = {
-                        id: `folder-${encodeURIComponent(folderPath)}`,
-                        name: part,
-                        path: folderPath,
-                        size: 0,
-                        level: partIndex,
-                        children: [],
-                        selected: true,
-                        partial: false,
-                    }
-                    folders.set(folderPath, folder)
-                    siblings.push(folder)
-                }
-                siblings = folder.children
-            })
-        })
-
-        this.refreshFolderState(roots)
-        return roots
-    }
-
-    private flattenFileTree(nodes: UploadFileTreeNode[]): UploadFileTreeNode[] {
-        return nodes.flatMap((node) => [node, ...this.flattenFileTree(node.children)])
-    }
-
-    private setNodeSelection(node: UploadFileTreeNode, selected: boolean) {
-        node.selected = selected
-        node.partial = false
-        node.children.forEach((child) => this.setNodeSelection(child, selected))
-        this.refreshFolderState(this.fileTree)
-    }
-
-    private refreshFolderState(nodes: UploadFileTreeNode[]) {
-        nodes.forEach((node) => {
-            if (node.children.length === 0) {
-                return
-            }
-
-            this.refreshFolderState(node.children)
-            node.size = node.children.reduce((size, child) => size + child.size, 0)
-            const selectedCount = node.children.filter((child) => child.selected && !child.partial).length
-            const partialCount = node.children.filter((child) => child.partial).length
-            node.selected = selectedCount === node.children.length && partialCount === 0
-            node.partial = (selectedCount > 0 || partialCount > 0) && !node.selected
-        })
-    }
-
-    private syncFileSelectionOptions() {
-        if (!this.hasFilesTab()) {
-            delete this.uploadOptions.fileSelection
-            return
-        }
-
-        const fileSelection = this.visibleFileTree
-            .filter((node) => node.fileIndex !== undefined)
-            .map((node): BittorrentFileSelection => ({
-                index: node.fileIndex!,
-                path: node.path,
-                name: node.name,
-                size: node.size,
-                wanted: node.selected,
-            }))
-
-        if (fileSelection.every((file) => file.wanted)) {
-            delete this.uploadOptions.fileSelection
-            return
-        }
-
-        this.uploadOptions.fileSelection = fileSelection
-    }
-
-    private updateIndeterminateCheckboxes() {
-        this.$timeout(() => {
-            this.visibleFileTree.forEach((node) => {
-                const element = document.getElementById(`upload-file-cb-${node.id}`) as HTMLInputElement | null
-                if (element) {
-                    element.indeterminate = node.partial
-                }
-            })
-        })
     }
 
 }
